test: reorganize tests by domain per ARCHITECTURE (Resolves #170)
Move 30+ misplaced tests under the
`tests/{unit,integration,infrastructure}/<domain>/...` layout from
docs/ARCHITECTURE.md and docs/TESTING.md, and split the god-class
`tests/test_security.py` into seven domain-scoped files.

Highlights:

* A. test_security.py split (7 classes -> 7 files), each new file's
  docstring records "Split from tests/test_security.py" for traceability.
* B/C. tests/integration/{,api/} flattening — 22 router/service/daemon
  files relocated under their owning domain, with the process/daemon/
  lifecycle set moved into tests/infrastructure/servers/ to match their
  existing `pytestmark = pytest.mark.slow`.
* D. tests/unit/{models,utils}/ dismantled — schedule models joined
  backups/, port_validation moved to servers/application/, config and
  database utility tests joined core/.
* E. tests/infrastructure/ — split aiohttp mocks, templates router
  smoke, parallel DB isolation, and performance monitoring tests into
  versions/templates/core subdirs.
* F. Empty tests/integration/api/, tests/unit/models/, and
  tests/unit/utils/ removed; new __init__.py files added to all
  integration and infrastructure subdirs to disambiguate the five new
  test_router.py modules pytest now collects in parallel.

Cross-cutting fixes:
* Update lazy tests.infrastructure.test_aiohttp_mocks imports in
  tests/unit/versions/application/test_{jar_cache_manager,version_manager}.py
  to the new path.
* Refresh docstring references in app/audit/application/legacy_facade.py,
  app/backups/adapters/legacy.py, app/servers/application/database_integration.py,
  and tests/unit/backups/test_legacy_shim.py so the test-file pointers
  still resolve after the moves.

Verification:
* uv run ruff check tests/ + uv run ruff format --check tests/ — same
  29 lint warnings / 5 format gaps as master baseline (no new findings).
* uv run pytest tests/unit -m "not slow" -> 1210 passed, 5 skipped.
* uv run pytest (full suite incl. slow) -> 1789 passed, 10 skipped.
* SKIP=pytest-unit-smoke,pytest-integration-smoke uv run pre-commit run
  --all-files -> all hooks passed.
