+++ tests/infrastructure/core/test_performance_monitoring.py
