+++ tests/infrastructure/servers/test_lifecycle.py
