+++ tests/infrastructure/templates/test_router_smoke.py
