+++ tests/integration/core/test_lifespan.py
