+++ tests/integration/core/visibility/test_router.py
