+++ tests/unit/core/test_config_validation.py
