+++ tests/unit/versions/application/test_version_manager.py
@@ -5,7 +5,7 @@ import pytest
 
 from app.servers.models import ServerType
 from app.versions.application.version_manager import MinecraftVersionManager, VersionInfo
-from tests.infrastructure.test_aiohttp_mocks import (
+from tests.infrastructure.versions.test_aiohttp_mocks import (
     MockAiohttpResponse,
     MockAiohttpSession,
 )
